+++ R/model-persistence.R
@@ -39,7 +39,6 @@ save_model_hdf5 <- function(object, filepath, overwrite = TRUE, include_optimize
                             filepath = filepath, 
                             overwrite = overwrite,
                             include_optimizer = include_optimizer)
-    mirror_to_run_dir(filepath)
     invisible(TRUE) 
   } else {
     invisible(FALSE)
@@ -103,7 +102,6 @@ save_model_weights_hdf5 <- function(object, filepath, overwrite = TRUE) {
   filepath <- normalize_path(filepath)
   if (confirm_overwrite(filepath, overwrite)) {
     object$save_weights(filepath = filepath, overwrite = overwrite)
-    mirror_to_run_dir(filepath)
     invisible(TRUE)
   } else {
     invisible(FALSE)
@@ -221,48 +219,52 @@ unserialize_model <- function(model, custom_objects = NULL, compile = TRUE) {
   load_model_hdf5(tmp, custom_objects = custom_objects, compile = compile)
 }
 
-
-# utility function to mirror saved models/weights into the run_dir
-# whenever a training_run is active
-mirror_to_run_dir <- function(filepath) {
-  mirror_path <- run_dir_path(filepath)
-  if (!is.null(mirror_path))
-    file.copy(filepath, mirror_path, overwrite = TRUE)
-}
-
-run_dir_path <- function(filepath, default = NULL) {
+model_to_tensors_info <- function(layers, name) {
+  named_layers <- lapply(layers, function(layer) {
+    tensorflow::tf$saved_model$utils$build_tensor_info(layer[[name]])
+  })
   
-  if (tfruns::is_run_active() && is_working_dir_path(filepath)) {
-    
-    # create the working dir path
-    run_dir_path <- file.path(run_dir(), filepath)
-    
-    # create directory if needed
-    target_dir <- dirname(run_dir_path)
-    if (!utils::file_test("-d", target_dir))
-      dir.create(target_dir, recursive = TRUE)
-    
-    # return the path
-    run_dir_path
-    
-  } else {
-    default
-  }
+  if (length(named_layers) == 1)
+    names(named_layers) <- name
+  else
+    names(named_layers) <- paste(name, seq_along(named_layers), sep = "")
   
+  named_layers
 }
 
-
-is_working_dir_path <- function(path) {
-  if (regexpr("^~", path) != -1L)
-    FALSE
-  else if (regexpr("^.:(/|\\\\)", path) != -1L)
-    FALSE
-  else if (regexpr("^(/|\\\\)", path) != -1L)
-    FALSE
-  else if (regexpr("^\\.\\.", path) != -1L)
-    FALSE
-  else 
-    TRUE
+#' Export a Saved Model
+#'
+#' Serialize a model to disk.
+#'
+#' @param object An \R object.
+#' @param export_dir_base A string containing a directory in which to create
+#'   versioned subdirectories containing exported SavedModels.
+#' 
+#' @return The path to the exported directory, as a string.
+#'
+#' @export
+export_savedmodel.keras.engine.training.Model <- function(object, export_dir_base) {
+  if (!is_backend("tensorflow"))
+    stop("'export_savedmodel' is only supported in the TensorFlow backend.")
+  
+  sess <- backend()$get_session()
+  
+  input_info <- model_to_tensors_info(object$input_layers, "input")
+  output_info <- model_to_tensors_info(object$output_layers, "output")
+  
+  builder <- tensorflow::tf$saved_model$builder$SavedModelBuilder(export_dir_base)
+  builder$add_meta_graph_and_variables(
+    sess,
+    list(
+      tensorflow::tf$python$saved_model$tag_constants$SERVING
+    ),
+    signature_def_map = list(
+      serving_default = tensorflow::tf$saved_model$signature_def_utils$build_signature_def(
+        inputs = input_info,
+        outputs = output_info
+      )
+    )
+  )
+  builder$save()
 }
-