+++ vignettes/examples/mnist_mlp.R
@@ -7,29 +7,30 @@
 
 library(keras)
 
+# Data Preparation ---------------------------------------------------
+
 batch_size <- 128
 num_classes <- 10
 epochs <- 30
 
-# the data, shuffled and split between train and test sets
-mnist <- dataset_mnist()
-x_train <- mnist$train$x
-y_train <- mnist$train$y
-x_test <- mnist$test$x
-y_test <- mnist$test$y
+# The data, shuffled and split between train and test sets
+c(c(x_train, y_train), c(x_test, y_test)) %<-% dataset_mnist()
 
-dim(x_train) <- c(nrow(x_train), 784)
-dim(x_test) <- c(nrow(x_test), 784)
+x_train <- array_reshape(x_train, c(nrow(x_train), 784))
+x_test <- array_reshape(x_test, c(nrow(x_test), 784))
 
+# Transform RGB values into [0,1] range
 x_train <- x_train / 255
 x_test <- x_test / 255
 
 cat(nrow(x_train), 'train samples\n')
 cat(nrow(x_test), 'test samples\n')
 
-# convert class vectors to binary class matrices
+# Convert class vectors to binary class matrices
 y_train <- to_categorical(y_train, num_classes)
 y_test <- to_categorical(y_test, num_classes)
+
+# Define Model --------------------------------------------------------------
 
 model <- keras_model_sequential()
 model %>% 
@@ -47,6 +48,9 @@ model %>% compile(
   metrics = c('accuracy')
 )
 
+# Training & Evaluation ----------------------------------------------------
+
+# Fit model to data
 history <- model %>% fit(
   x_train, y_train,
   batch_size = batch_size,
@@ -61,7 +65,8 @@ score <- model %>% evaluate(
   x_test, y_test,
   verbose = 0
 )
-  
+
+# Output metrics
 cat('Test loss:', score[[1]], '\n')
 cat('Test accuracy:', score[[2]], '\n')