+++ website/articles/examples/imdb_bidirectional_lstm.R
@@ -5,44 +5,59 @@
 
 library(keras)
 
+# Define maximum number of input features
 max_features <- 20000
 
-# cut texts after this number of words
+# Cut texts after this number of words
 # (among top max_features most common words)
 maxlen <- 100
 
 batch_size <- 32
 
+# Load imdb dataset 
 cat('Loading data...\n')
 imdb <- dataset_imdb(num_words = max_features)
+
+# Define training and test sets
 x_train <- imdb$train$x
 y_train <- imdb$train$y
 x_test <- imdb$test$x
 y_test <- imdb$test$y
 
+# Output lengths of testing and training sets
 cat(length(x_train), 'train sequences\n')
 cat(length(x_test), 'test sequences\n')
 
 cat('Pad sequences (samples x time)\n')
+
+# Pad training and test inputs
 x_train <- pad_sequences(x_train, maxlen = maxlen)
 x_test <- pad_sequences(x_test, maxlen = maxlen)
+
+# Output dimensions of training and test inputs
 cat('x_train shape:', dim(x_train), '\n')
 cat('x_test shape:', dim(x_test), '\n')
 
+# Initialize model
 model <- keras_model_sequential()
 model %>%
-  layer_embedding(input_dim = max_features, output_dim = 128, input_length = maxlen) %>% 
-  bidirectional(layer_lstm(units = 64)) %>% 
+  # Creates dense embedding layer; outputs 3D tensor
+  # with shape (batch_size, sequence_length, output_dim)
+  layer_embedding(input_dim = max_features, 
+                  output_dim = 128, 
+                  input_length = maxlen) %>% 
+  bidirectional(layer_lstm(units = 64)) %>%
   layer_dropout(rate = 0.5) %>% 
   layer_dense(units = 1, activation = 'sigmoid')
 
-# try using different optimizers and different optimizer configs
+# Try using different optimizers and different optimizer configs
 model %>% compile(
   loss = 'binary_crossentropy',
   optimizer = 'adam',
   metrics = c('accuracy')
 )
 
+# Train model over four epochs
 cat('Train...\n')
 model %>% fit(
   x_train, y_train,
@@ -50,5 +65,3 @@ model %>% fit(
   epochs = 4,
   validation_data = list(x_test, y_test)
 )
-
-