+++ website/articles/examples/mnist_acgan.R
@@ -6,11 +6,11 @@
 #' very slow on the CPU. Prefer the TensorFlow backend if you plan on iterating,
 #' as the compilation time can be a blocker using Theano.  
 #'   
-#' Hardware           | Backend | Time / Epoch
-#' | -----------------| ------- | ------------------- |
-#' |   CPU               | TF       | 3 hrs | 
-#' |   Titan X (maxwell) | TF       | 4 min |
-#' |   Titan X (maxwell) | TH       | 7 min |
+#' | Hardware         | Backend | Time / Epoch        |
+#' | ---------------- | ------- | ------------------- |
+#' |CPU               | TF      | 3 hrs               | 
+#' |Titan X (maxwell) | TF      | 4 min               |
+#' |Titan X (maxwell) | TH      | 7 min               |
 #' 
 
 library(keras)
@@ -23,7 +23,7 @@ K$set_image_data_format('channels_first')
 
 build_generator <- function(latent_size){
   
-  # we will map a pair of (z, L), where z is a latent vector and L is a
+  # We will map a pair of (z, L), where z is a latent vector and L is a
   # label drawn from P_c, to image space (..., 1, 28, 28)
   cnn <- keras_model_sequential()
   
@@ -31,29 +31,29 @@ build_generator <- function(latent_size){
     layer_dense(1024, input_shape = latent_size, activation = "relu") %>%
     layer_dense(128*7*7, activation = "relu") %>%
     layer_reshape(c(128, 7, 7)) %>%
-    # upsample to (..., 14, 14)
+    # Upsample to (..., 14, 14)
     layer_upsampling_2d(size = c(2, 2)) %>%
     layer_conv_2d(
       256, c(5,5), padding = "same", activation = "relu",
       kernel_initializer = "glorot_normal"
     ) %>%
-    # upsample to (..., 28, 28)
+    # Upsample to (..., 28, 28)
     layer_upsampling_2d(size = c(2, 2)) %>%
     layer_conv_2d(
       128, c(5,5), padding = "same", activation = "tanh",
       kernel_initializer = "glorot_normal"
     ) %>%
-    # take a channel axis reduction
+    # Take a channel axis reduction
     layer_conv_2d(
       1, c(2,2), padding = "same", activation = "tanh",
       kernel_initializer = "glorot_normal"
     )
   
   
-  # this is the z space commonly refered to in GAN papers
+  # This is the z space commonly refered to in GAN papers
   latent <- layer_input(shape = list(latent_size))
   
-  # this will be our label
+  # This will be our label
   image_class <- layer_input(shape = list(1))
   
   # 10 classes in MNIST
@@ -65,7 +65,7 @@ build_generator <- function(latent_size){
     layer_flatten()
   
   
-  # hadamard product between z-space and a class conditional embedding
+  # Hadamard product between z-space and a class conditional embedding
   h <- layer_multiply(list(latent, cls))
   
   fake_image <- cnn(h)
@@ -75,7 +75,7 @@ build_generator <- function(latent_size){
 
 build_discriminator <- function(){
   
-  # build a relatively standard conv net, with LeakyReLUs as suggested in
+  # Build a relatively standard conv net, with LeakyReLUs as suggested in
   # the reference paper
   cnn <- keras_model_sequential()
   
@@ -106,7 +106,7 @@ build_discriminator <- function(){
   image <- layer_input(shape = c(1, 28, 28))
   features <- cnn(image)
   
-  # first output (name=generation) is whether or not the discriminator
+  # First output (name=generation) is whether or not the discriminator
   # thinks the image that is being shown is fake, and the second output
   # (name=auxiliary) is the class that the discriminator thinks the image
   # belongs to.
@@ -121,7 +121,7 @@ build_discriminator <- function(){
 
 # Parameters --------------------------------------------------------------
 
-# batch and latent size taken from the paper
+# Batch and latent size taken from the paper
 epochs <- 50
 batch_size <- 100
 latent_size <- 100
@@ -130,16 +130,16 @@ latent_size <- 100
 adam_lr <- 0.00005 
 adam_beta_1 <- 0.5
 
-# Model definition --------------------------------------------------------
+# Model Definition --------------------------------------------------------
 
-# build the discriminator
+# Build the discriminator
 discriminator <- build_discriminator()
 discriminator %>% compile(
   optimizer = optimizer_adam(lr = adam_lr, beta_1 = adam_beta_1),
   loss = list("binary_crossentropy", "sparse_categorical_crossentropy")
 )
 
-# build the generator
+# Build the generator
 generator <- build_generator(latent_size)
 generator %>% compile(
   optimizer = optimizer_adam(lr = adam_lr, beta_1 = adam_beta_1),
@@ -151,9 +151,8 @@ image_class <- layer_input(shape = list(1), dtype = "int32")
 
 fake <- generator(list(latent, image_class))
 
-# we only want to be able to train generation for the combined model
-
-discriminator$trainable <- FALSE
+# Only want to be able to train generation for the combined model
+freeze_layers(discriminator)
 results <- discriminator(fake)
 
 combined <- keras_model(list(latent, image_class), results)
@@ -162,16 +161,15 @@ combined %>% compile(
   loss = list("binary_crossentropy", "sparse_categorical_crossentropy")
 )
 
+# Data Preparation --------------------------------------------------------
 
-# Data preparation --------------------------------------------------------
-
-# get our mnist data, and force it to be of shape (..., 1, 28, 28) with
+# Loade mnist data, and force it to be of shape (..., 1, 28, 28) with
 # range [-1, 1]
 mnist <- dataset_mnist()
 mnist$train$x <- (mnist$train$x - 127.5)/127.5
 mnist$test$x <- (mnist$test$x - 127.5)/127.5
-dim(mnist$train$x) <- c(60000, 1, 28, 28) 
-dim(mnist$test$x) <- c(10000, 1, 28, 28) 
+mnist$train$x <- array_reshape(mnist$train$x, c(60000, 1, 28, 28))
+mnist$test$x <- array_reshape(mnist$test$x, c(10000, 1, 28, 28))
 
 num_train <- dim(mnist$train$x)[1]
 num_test <- dim(mnist$test$x)[1]
@@ -196,21 +194,21 @@ for(epoch in 1:epochs){
     
     pb$tick()
     
-    # generate a new batch of noise
+    # Generate a new batch of noise
     noise <- runif(n = batch_size*latent_size, min = -1, max = 1) %>%
       matrix(nrow = batch_size, ncol = latent_size)
     
-    # get a batch of real images
+    # Get a batch of real images
     batch <- sample(possible_indexes, size = batch_size)
     possible_indexes <- possible_indexes[!possible_indexes %in% batch]
     image_batch <- mnist$train$x[batch,,,,drop = FALSE]
     label_batch <- mnist$train$y[batch]
     
-    # sample some labels from p_c
+    # Sample some labels from p_c
     sampled_labels <- sample(0:9, batch_size, replace = TRUE) %>%
       matrix(ncol = 1)
     
-    # generate a batch of fake images, using the generated labels as a
+    # Generate a batch of fake images, using the generated labels as a
     # conditioner. We reshape the sampled labels to be
     # (batch_size, 1) so that we can feed them into the embedding
     # layer as a length one sequence
@@ -220,7 +218,7 @@ for(epoch in 1:epochs){
     y <- c(rep(1L, batch_size), rep(0L, batch_size)) %>% matrix(ncol = 1)
     aux_y <- c(label_batch, sampled_labels) %>% matrix(ncol = 1)
     
-    # see if the discriminator can figure itself out...
+    # Check if the discriminator can figure itself out
     disc_loss <- train_on_batch(
       discriminator, x = X, 
       y = list(y, aux_y)
@@ -228,15 +226,15 @@ for(epoch in 1:epochs){
     
     epoch_disc_loss <- rbind(epoch_disc_loss, unlist(disc_loss))
     
-    # make new noise. we generate 2 * batch size here such that we have
-    # the generator optimize over an identical number of images as the
+    # Make new noise. Generate 2 * batch size here such that
+    # the generator optimizes over an identical number of images as the
     # discriminator
     noise <- runif(2*batch_size*latent_size, min = -1, max = 1) %>%
       matrix(nrow = 2*batch_size, ncol = latent_size)
     sampled_labels <- sample(0:9, size = 2*batch_size, replace = TRUE) %>%
       matrix(ncol = 1)
     
-    # we want to train the generator to trick the discriminator
+    # Want to train the generator to trick the discriminator
     # For the generator, we want all the {fake, not-fake} labels to say
     # not-fake
     trick <- rep(1, 2*batch_size) %>% matrix(ncol = 1)
@@ -253,13 +251,13 @@ for(epoch in 1:epochs){
   
   cat(sprintf("\nTesting for epoch %02d:", epoch))
   
-  # evaluate the testing loss here
+  # Evaluate the testing loss here
   
-  # generate a new batch of noise
+  # Generate a new batch of noise
   noise <- runif(num_test*latent_size, min = -1, max = 1) %>%
     matrix(nrow = num_test, ncol = latent_size)
   
-  # sample some labels from p_c and generate images from them
+  # Sample some labels from p_c and generate images from them
   sampled_labels <- sample(0:9, size = num_test, replace = TRUE) %>%
     matrix(ncol = 1)
   generated_images <- predict(generator, list(noise, sampled_labels))
@@ -268,7 +266,7 @@ for(epoch in 1:epochs){
   y <- c(rep(1, num_test), rep(0, num_test)) %>% matrix(ncol = 1)
   aux_y <- c(mnist$test$y, sampled_labels) %>% matrix(ncol = 1)
   
-  # see if the discriminator can figure itself out...
+  # See if the discriminator can figure itself out...
   discriminator_test_loss <- evaluate(
     discriminator, X, list(y, aux_y), 
     verbose = FALSE
@@ -276,7 +274,7 @@ for(epoch in 1:epochs){
   
   discriminator_train_loss <- apply(epoch_disc_loss, 2, mean)
   
-  # make new noise
+  # Make new noise
   noise <- runif(2*num_test*latent_size, min = -1, max = 1) %>%
     matrix(nrow = 2*num_test, ncol = latent_size)
   sampled_labels <- sample(0:9, size = 2*num_test, replace = TRUE) %>%
@@ -293,7 +291,7 @@ for(epoch in 1:epochs){
   generator_train_loss <- apply(epoch_gen_loss, 2, mean)
   
   
-  # generate an epoch report on performance
+  # Generate an epoch report on performance
   row_fmt <- "\n%22s : loss %4.2f | %5.2f | %5.2f"
   cat(sprintf(
     row_fmt, 
@@ -317,6 +315,7 @@ for(epoch in 1:epochs){
     discriminator_train_loss[2],
     discriminator_train_loss[3]
   ))
+  
   cat(sprintf(
     row_fmt, 
     "discriminator (test)",
@@ -327,14 +326,14 @@ for(epoch in 1:epochs){
   
   cat("\n")
   
-  # generate some digits to display
+  # Generate some digits to display
   noise <- runif(10*latent_size, min = -1, max = 1) %>%
     matrix(nrow = 10, ncol = latent_size)
   
   sampled_labels <- 0:9 %>%
     matrix(ncol = 1)
   
-  # get a batch to display
+  # Get a batch to display
   generated_images <- predict(
     generator,    
     list(noise, sampled_labels)