+++ website/articles/examples/mnist_antirectifier.R
@@ -5,48 +5,48 @@
 #' `compute_output_shape` and `call`.
 #'
 #' Note that the same result can also be achieved via a Lambda layer.
-#' 
-
-#' ## Data Preparation
+#'
 
 library(keras)
+
+# Data Preparation --------------------------------------------------------
 
 batch_size <- 128
 num_classes <- 10
 epochs <- 40
 
-# the data, shuffled and split between train and test sets
+# The data, shuffled and split between train and test sets
 mnist <- dataset_mnist()
 x_train <- mnist$train$x
 y_train <- mnist$train$y
 x_test <- mnist$test$x
 y_test <- mnist$test$y
 
-dim(x_train) <- c(nrow(x_train), 784)
-dim(x_test) <- c(nrow(x_test), 784)
+# Redimension
+x_train <- array_reshape(x_train, c(nrow(x_train), 784))
+x_test <- array_reshape(x_test, c(nrow(x_test), 784))
 
+# Transform RGB values into [0,1] range
 x_train <- x_train / 255
 x_test <- x_test / 255
 
 cat(nrow(x_train), 'train samples\n')
 cat(nrow(x_test), 'test samples\n')
 
-# convert class vectors to binary class matrices
+# Convert class vectors to binary class matrices
 y_train <- to_categorical(y_train, num_classes)
 y_test <- to_categorical(y_test, num_classes)
 
-#' ## Antirectifier Layer
+# Antirectifier Layer -----------------------------------------------------
 #'
-#' This is the combination of a sample-wise L2 normalization
-#' with the concatenation of the positive part of the input with the negative
-#' part of the input. The result is a tensor of samples that are twice as large
+#' This is the combination of a sample-wise L2 normalization with the
+#' concatenation of the positive part of the input with the negative part
+#' of the input. The result is a tensor of samples that are twice as large
 #' as the input samples.
 #'
 #' It can be used in place of a ReLU.
-#'
-#' Input shape: 2D tensor of shape (samples, n)
-#'
-#' Output shape: 2D tensor of shape (samples, 2*n)
+#'  Input shape: 2D tensor of shape (samples, n)
+#'  Output shape: 2D tensor of shape (samples, 2*n)
 #'
 #' When applying ReLU, assuming that the distribution of the previous output is
 #' approximately centered around 0., you are discarding half of your input. This
@@ -55,12 +55,13 @@ y_test <- to_categorical(y_test, num_classes)
 #' Antirectifier allows to return all-positive outputs like ReLU, without
 #' discarding any data.
 #'
-#' Tests on MNIST show that Antirectifier allows to train networks with twice
-#' less parameters yet with comparable classification accuracy as an equivalent
+#' Tests on MNIST show that Antirectifier allows to train networks with half
+#' the parameters yet with comparable classification accuracy as an equivalent
 #' ReLU-based network.
+#'
+#' Because our custom layer is written with primitives from the Keras backend
+#' (`K`), our code can run both on TensorFlow and Theano.
 
-# Because our custom layer is written with primitives from the Keras backend
-# (`K`), our code can run both on TensorFlow and Theano.
 K <- backend()
 
 # Custom layer class
@@ -86,13 +87,13 @@ AntirectifierLayer <- R6::R6Class("KerasLayer",
   )
 )
 
-# create layer wrapper function
+# Create layer wrapper function
 layer_antirectifier <- function(object) {
   create_layer(AntirectifierLayer, object)
 }
 
 
-#' ## Define and Train Model
+# Define & Train Model -------------------------------------------------
 
 model <- keras_model_sequential()
 model %>% 
@@ -104,18 +105,17 @@ model %>%
   layer_dropout(rate = 0.1) %>%
   layer_dense(units = 10, activation = 'softmax')
 
-# compile the model
+# Compile the model
 model %>% compile(
   loss = 'categorical_crossentropy',
   optimizer = 'rmsprop',
   metrics = c('accuracy')
 )
 
-# train the model
+# Train the model
 model %>% fit(x_train, y_train,
   batch_size = batch_size,
   epochs = epochs,
   verbose = 1,
   validation_data= list(x_test, y_test)
 )
-